added extra subscriber node for interpreting street sign data
